content/html/content/test/forms/test_option_text.html

<!doctype html>
<meta charset=utf-8>
<title>HTMLOptionElement.text</title>
<link rel=author title=Ms2ger href="mailto:Ms2ger@gmail.com">
<link rel=help href="http://www.whatwg.org/html/#dom-option-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var option = document.createElement("option");
  option.appendChild(document.createElement("font"))
        .appendChild(document.createTextNode(" font "));
  assert_equals(option.text, "font");
});
</script>
